perf(ring-switch): peel block axis in m-eval
Strip the power-of-two num_blocks axis out of the separable w and t terms
so batched verifier paths can keep using a succinct eq-weighted evaluation
even when the outer claim dimensions are ragged.

Also add the shared offset_eq helper for 2-adic peeled carry summaries and
clean up the remaining clippy issues in the sibling worktree.

Made-with: Cursor

+++ src/algebra/fields/fp128.rs
@@ -724,7 +724,7 @@ impl<const P: u128> Fp128<P> {
                 b1 = in(reg) b[1],
                 add_lo = in(reg) addend[0],
                 add_hi = in(reg) addend[1],
-                c = in(reg) Self::C_LO as u64,
+                c = in(reg) Self::C_LO,
                 p00l = out(reg) _,
                 p00h = out(reg) _,
                 p01l = out(reg) _,
